mechanical: moved function more appropriate files
