+++ repo/object/object_manager.go
@@ -2,9 +2,7 @@
 package object
 
 import (
-	"bytes"
 	"context"
-	"encoding/json"
 	"io"
 
 	"github.com/pkg/errors"
@@ -77,11 +75,6 @@ func (om *Manager) NewWriter(ctx context.Context, opt WriterOptions) Writer {
 	w.initBuffer()
 
 	return w
-}
-
-// Open creates new ObjectReader for reading given object from a repository.
-func Open(ctx context.Context, r contentReader, objectID ID) (Reader, error) {
-	return openAndAssertLength(ctx, r, objectID, -1)
 }
 
 // Concatenate creates an object that's a result of concatenation of other objects. This is more efficient than reading
@@ -183,77 +176,6 @@ func appendIndexEntries(indexEntries []indirectObjectEntry, startingLength int64
 	return indexEntries, totalLength
 }
 
-func openAndAssertLength(ctx context.Context, cr contentReader, objectID ID, assertLength int64) (Reader, error) {
-	if indexObjectID, ok := objectID.IndexObjectID(); ok {
-		// recursively calls openAndAssertLength
-		seekTable, err := loadSeekTable(ctx, cr, indexObjectID)
-		if err != nil {
-			return nil, err
-		}
-
-		totalLength := seekTable[len(seekTable)-1].endOffset()
-
-		return &objectReader{
-			ctx:         ctx,
-			cr:          cr,
-			seekTable:   seekTable,
-			totalLength: totalLength,
-		}, nil
-	}
-
-	return newRawReader(ctx, cr, objectID, assertLength)
-}
-
-// VerifyObject ensures that all objects backing ObjectID are present in the repository
-// and returns the content IDs of which it is composed.
-func VerifyObject(ctx context.Context, cr contentReader, oid ID) ([]content.ID, error) {
-	tracker := &contentIDTracker{}
-
-	if err := verifyObjectInternal(ctx, cr, oid, tracker); err != nil {
-		return nil, err
-	}
-
-	return tracker.contentIDs(), nil
-}
-
-func verifyIndirectObjectInternal(ctx context.Context, cr contentReader, indexObjectID ID, tracker *contentIDTracker) error {
-	if err := verifyObjectInternal(ctx, cr, indexObjectID, tracker); err != nil {
-		return errors.Wrap(err, "unable to read index")
-	}
-
-	seekTable, err := loadSeekTable(ctx, cr, indexObjectID)
-	if err != nil {
-		return err
-	}
-
-	for _, m := range seekTable {
-		err := verifyObjectInternal(ctx, cr, m.Object, tracker)
-		if err != nil {
-			return err
-		}
-	}
-
-	return nil
-}
-
-func verifyObjectInternal(ctx context.Context, r contentReader, oid ID, tracker *contentIDTracker) error {
-	if indexObjectID, ok := oid.IndexObjectID(); ok {
-		return verifyIndirectObjectInternal(ctx, r, indexObjectID, tracker)
-	}
-
-	if contentID, _, ok := oid.ContentID(); ok {
-		if _, err := r.ContentInfo(ctx, contentID); err != nil {
-			return errors.Wrapf(err, "error getting content info for %v", contentID)
-		}
-
-		tracker.addContentID(contentID)
-
-		return nil
-	}
-
-	return errors.Errorf("unrecognized object type: %v", oid)
-}
-
 func nullTrace(message string, args ...interface{}) {
 }
 
@@ -297,103 +219,4 @@ func NewObjectManager(ctx context.Context, bm contentManager, f Format, opts Man
 func (om *Manager) Close() error {
 	om.bufferPool.Close()
 	return nil
-}
-
-/*
-
-{"stream":"kopia:indirect","entries":[
-{"l":1698099,"o":"D13ea27f9ad891ad4a2edfa983906863d"},
-{"s":1698099,"l":1302081,"o":"De8ca8327cd3af5f4edbd5ed1009c525e"},
-{"s":3000180,"l":4352499,"o":"D6b6eb48ca5361d06d72fe193813e42e1"},
-{"s":7352679,"l":1170821,"o":"Dd14653f76b63802ed48be64a0e67fea9"},
-
-{"s":91094118,"l":1645153,"o":"Daa55df764d881a1daadb5ea9de17abbb"}
-]}
-*/
-
-type indirectObject struct {
-	StreamID string                `json:"stream"`
-	Entries  []indirectObjectEntry `json:"entries"`
-}
-
-func loadSeekTable(ctx context.Context, cr contentReader, indexObjectID ID) ([]indirectObjectEntry, error) {
-	r, err := openAndAssertLength(ctx, cr, indexObjectID, -1)
-	if err != nil {
-		return nil, err
-	}
-	defer r.Close() //nolint:errcheck
-
-	var ind indirectObject
-
-	if err := json.NewDecoder(r).Decode(&ind); err != nil {
-		return nil, errors.Wrap(err, "invalid indirect object")
-	}
-
-	return ind.Entries, nil
-}
-
-func newRawReader(ctx context.Context, cr contentReader, objectID ID, assertLength int64) (Reader, error) {
-	contentID, compressed, ok := objectID.ContentID()
-	if !ok {
-		return nil, errors.Errorf("unsupported object ID: %v", objectID)
-	}
-
-	payload, err := cr.GetContent(ctx, contentID)
-	if errors.Is(err, content.ErrContentNotFound) {
-		return nil, errors.Wrapf(ErrObjectNotFound, "content %v not found", contentID)
-	}
-
-	if err != nil {
-		return nil, errors.Wrap(err, "unexpected content error")
-	}
-
-	if compressed {
-		var b bytes.Buffer
-
-		if err = decompress(&b, payload); err != nil {
-			return nil, errors.Wrap(err, "decompression error")
-		}
-
-		payload = b.Bytes()
-	}
-
-	if assertLength != -1 && int64(len(payload)) != assertLength {
-		return nil, errors.Errorf("unexpected chunk length %v, expected %v", len(payload), assertLength)
-	}
-
-	return newObjectReaderWithData(payload), nil
-}
-
-func decompress(output *bytes.Buffer, b []byte) error {
-	compressorID, err := compression.IDFromHeader(b)
-	if err != nil {
-		return errors.Wrap(err, "invalid compression header")
-	}
-
-	compressor := compression.ByHeaderID[compressorID]
-	if compressor == nil {
-		return errors.Errorf("unsupported compressor %x", compressorID)
-	}
-
-	return compressor.Decompress(output, b)
-}
-
-type readerWithData struct {
-	io.ReadSeeker
-	length int64
-}
-
-func (rwd *readerWithData) Close() error {
-	return nil
-}
-
-func (rwd *readerWithData) Length() int64 {
-	return rwd.length
-}
-
-func newObjectReaderWithData(data []byte) Reader {
-	return &readerWithData{
-		ReadSeeker: bytes.NewReader(data),
-		length:     int64(len(data)),
-	}
 }
